requirements.txt
Flask
kubernetes
requests
python-dotenv
Werkzeug